requirements.txt
amqp==5.1.1
async-timeout==4.0.2
bcrypt==4.0.1
billiard==3.6.4.0
cachelib==0.9.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.6
cycler==0.11.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
fonttools==4.38.0
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.24.0
packaging==22.0
pdfkit==1.0.0
Pillow==9.3.0
prompt-toolkit==3.0.36
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
redis==4.4.0
redmail==0.4.2
six==1.16.0
SQLAlchemy==1.4.45
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
